requirements.txt
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=0.21.0
rich>=13.0.0
prompt-toolkit>=3.0.0
httpx>=0.24.0
pyte>=0.8.0
